model/model/code/requirements.txt
torch
torchvision
Pillow
boto3
